+++ catalog/src/main/java/it/eng/catalog/rest/api/ProxyAPIController.java
@@ -4,8 +4,8 @@ import java.util.List;
 
 import org.springframework.http.MediaType;
 import org.springframework.http.ResponseEntity;
-import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.PathVariable;
+import org.springframework.web.bind.annotation.PostMapping;
 import org.springframework.web.bind.annotation.RequestBody;
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RestController;
@@ -13,14 +13,15 @@ import org.springframework.web.bind.annotation.RestController;
 import com.fasterxml.jackson.databind.JsonNode;
 
 import it.eng.catalog.model.Catalog;
+import it.eng.catalog.serializer.Serializer;
 import it.eng.catalog.service.ProxyAPIService;
 import it.eng.tools.controller.ApiEndpoints;
-import it.eng.tools.model.Serializer;
 import it.eng.tools.response.GenericApiResponse;
 import lombok.extern.slf4j.Slf4j;
 
 @RestController
-@RequestMapping(consumes = MediaType.APPLICATION_JSON_VALUE, produces = MediaType.APPLICATION_JSON_VALUE, path = ApiEndpoints.PROXY_V1)
+@RequestMapping(consumes = MediaType.APPLICATION_JSON_VALUE, produces = MediaType.APPLICATION_JSON_VALUE, 
+	path = ApiEndpoints.PROXY_V1)
 @Slf4j
 public class ProxyAPIController {
 
@@ -31,7 +32,7 @@ public class ProxyAPIController {
 		this.proxyApiService = proxyApiService;
 	}
 
-	@GetMapping(path = "/datasets/{id}/formats")
+	@PostMapping(path = "/datasets/{id}/formats")
 	public ResponseEntity<GenericApiResponse<List<String>>> getFormatsFromDataset(@PathVariable String id,
 			@RequestBody JsonNode formatsRequest) {
 		log.info("Fetching formats from dataset with id: '" + id + "'");
@@ -42,7 +43,7 @@ public class ProxyAPIController {
 				.body(GenericApiResponse.success(formats, "Fetched formats"));
 	}
 	
-	@GetMapping(path = "/catalogs")
+	@PostMapping(path = "/catalogs")
 	public ResponseEntity<GenericApiResponse<JsonNode>> getCatalog(@RequestBody JsonNode formatsRequest) {
 		log.info("Fetching proxy catalog");
 		String forwardTo = formatsRequest.get("Forward-To").asText();